% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA.utilities.R
\name{convert.char}
\alias{convert.char}
\title{Convert character}
\usage{
convert.char(character)
}
\arguments{
\item{character}{List, numeric, character or matrix describing a morphological character}
}
\value{
The character as a list, with each entry relating the state of the character
for the corresponding taxon
}
\description{
Converts a character (inapplicable or missing)
}
\keyword{internal}
