% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA.reconstruction.R
\name{first.downpass}
\alias{first.downpass}
\title{First downpass}
\usage{
first.downpass(states_matrix)
}
\arguments{
\item{states_matrix}{A \code{list} contains all the states and the activations}
}
\description{
Applies a first down pass to a node
}
\examples{
## Create a states matrix for reconstruction
tree <- read.tree(text = "((a,b),(c,d));")
NA_matrix <- make.states.matrix(tree, "01?-")

## Apply the first downpass
(NA_matrix <- first.downpass(NA_matrix))

## Access the states
NA_matrix$Dp1

}
\author{
Thomas Guillerme
}

