% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fitch.reconstruction.R
\name{fitch.downpass}
\alias{fitch.downpass}
\title{Fitch downpass}
\usage{
fitch.downpass(states_matrix, tree)
}
\arguments{
\item{states_matrix}{A \code{list} contains all the states and the activations}

\item{tree}{A \code{phylo} tree}
}
\description{
Applies a Fitch down pass to a node
}
\author{
Thomas Guillerme
}

