% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.algorithm.R
\name{generate.palette}
\alias{generate.palette}
\title{Generate palette for tips}
\usage{
generate.palette(labels)
}
\arguments{
\item{labels}{Characters, for example from c(0:9, '-'), labelling the state
of each tip.}
}
\description{
Generate palette for tips
}
\keyword{internal}
