% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inapp-package.R
\docType{package}
\name{Inapp-package}
\alias{Inapp-package}
\alias{Inapp}
\title{Inapplicable Discrete Characters Reconstruction}
\description{
An algorithm allowing states reconstruction and tree length counting for inapplicable discrete characters
}
\examples{

}
\references{
Smith, M. R., Brazeau, M. D., and Guillerme, T., 20??. @@
}
\author{
Thomas Guillerme <guillert@tcd.ie>
}
\keyword{characters,}
\keyword{inapplicable}
\keyword{length}
\keyword{maximum}
\keyword{parsimony,}
\keyword{reconstruction,}
\keyword{states}
\keyword{tree}
