% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg.R
\name{MatrixData}
\alias{MatrixData}
\title{Matrix data}
\usage{
MatrixData(states_matrix, fitch_states, state.labels)
}
\arguments{
\item{states_matrix}{The output of \code{apply.reconstruction}}

\item{fitch_states}{The output of \code{apply.reconstruction(method='Fitch')}}

\item{state.labels}{Character vector translating tokens into characer states,
e.g. \code{c('0: Absent', '1: Present')}}
}
\value{
A list summarising details of the matrix, with named entries;
\code{edge_col} (edge colours); \code{edge_col_array} (formatted as a Javascript array),
\code{legend} (detailing what colours mean); \code{legend_col} (colour for legend entries);
\code{tips_labels}; \code{tips_colours}; \code{dud_steps} (Steps reconstructed by Fitch,
but not the inapplicable algorithm)
}
\description{
Generates and summarises data from a \code{states_matrix} that has been output
by \code{apply.reconstruction}.
}
\author{
Martin R. Smith
}
\keyword{internal}
