% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{palettes}
\alias{palettes}
\title{Palettes}
\format{An object of class \code{list} of length 11.}
\source{
\href{http://colorbrewer2.org/#type=diverging&scheme=RdYlBu&n=3}{ColourBrewer2.org}
}
\usage{
palettes
}
\description{
A list of eleven Brewer palettes containing one to eleven colours that
are readily distinguished by colourblind viewers
}
\keyword{datasets}
