% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.algorithm.R
\name{plot.states.matrix}
\alias{plot.states.matrix}
\title{Plot reconstruction}
\usage{
\method{plot}{states.matrix}(states_matrix, tree, passes = c(1, 2, 3, 4),
  show.labels = 0, col.tips.nodes = c("orange", "bisque2", "lightblue"),
  counts = 0, ...)
}
\arguments{
\item{states_matrix}{\code{list}, a list of state changes from \code{inapplicable.algorithm}}

\item{tree}{\code{phylo}, the tree used in the analysis}

\item{passes}{\code{numeric}, the number of passes to plot (default = \code{c(1,2,3,4)})}

\item{show.labels}{\code{numeric}, either \code{1} for showing the tip labels, \code{2} for the node labels or \code{c(1,2)} for both (default = \code{NULL}).}

\item{col.tips.nodes}{\code{character}, a vector of up to three colors to be used for displaying respectively the tips, the nodes and the activated/counted nodes (if \code{counts != 0}).}

\item{counts}{\code{numeric}, whether to display the activations (\code{1}) or/and the homoplasies (\code{2}) or nothing (\code{0}; default).}

\item{...}{any optional arguments to be passed to \code{\link[ape]{plot.phylo}}}
}
\description{
Plots an ancestral states reconstruction and tree length
}
\author{
Thomas Guillerme
}

