% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg.R
\name{PlotCharacterMapping}
\alias{PlotCharacterMapping}
\title{Plot character mapping}
\usage{
PlotCharacterMapping(char, stateLabels, singleTree, legendText,
  SetPar = par(mar = rep(0.2, 4), cex = 0.7), canvas, treeNames,
  analysisLabels = svgCanvas$analysisNames, svgFilename)
}
\arguments{
\item{char}{Character vector specifying distribution of tokens among (named) taxa}

\item{stateLabels}{Character vector specifying the labels to apply to each state}

\item{singleTree}{A single tree to be plotted in Latex output}

\item{legendText}{Character giving legend text to print in Latex output}

\item{SetPar}{Graphical parameters to set before plotting PNG tree for Latex output}

\item{canvas}{an object of class SVGForest.}

\item{treeNames}{Character string naming the analysis used to generate each tree}

\item{analysisLabels}{Characte vector specifying names of each analysis, in order to be printed}

\item{svgFilename}{Character string specifying location to save each file,
containing the expression \code{%s}, which will be replaced with the number of the tree.}
}
\value{
Prints the tree in an appropriate markdown format
}
\description{
Plot a tree with a character reconstruction plotted upon it
}
\author{
Martin R. Smith
}
