% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg.R
\name{PrintJavascript}
\alias{PrintJavascript}
\title{Print Javascript}
\usage{
PrintJavascript(filepath)
}
\arguments{
\item{filepath}{Path to Javascript template}
}
\value{
Prints the contents of the Javascript file to to stdout,
executing any R code included in r blocks in the file.
}
\description{
Allows arbitrary JavaScript content to be imported into an HTML document
produced with \code{bookdown} or \code{rmarkdown}.  Suggested for use in Rmd documents
that will be rendered in HTML; useful in supoprting user interaction with
SVG elements produced by \link{PlotCharacterMapping}.
}
\details{
Note that the content will be ignored when files are compiled for PDF output.
}
\author{
Martin R. Smith
}
