% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runInapp.R
\name{runInapp}
\alias{runInapp}
\title{Run Inapp}
\usage{
runInapp(..., remote = TRUE)
}
\arguments{
\item{...}{Any optional option to be passed to shiny::runApp}

\item{remote}{Whether to run the app from a remote server (\code{TRUE}, default - \code{shiny::runGitHub("Inapp", "TGuillerme", ...)}) or from the local server (\code{FALSE} - \code{shiny::runApp(".", ...)}).}
}
\description{
Running the Inapp GUI via shiny
}
\examples{
\dontrun{
## Running the GUI on your web browser
runInapp()
}

}
\seealso{
\code{\link{apply.reconstruction}}, \code{\link{plot.states.matrix}}
}
\author{
Thomas Guillerme
}
