% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA.utilities.R
\name{score.from}
\alias{score.from}
\title{Score from marked nodes}
\usage{
score.from(counted.nodes)
}
\arguments{
\item{counted.nodes}{An \code{integer} vector listing nodes at which a score was noted}
}
\value{
The number of items in the vector: zero if none
}
\description{
Score from marked nodes
}
\keyword{internal}
