% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA.reconstruction.R
\name{second.downpass}
\alias{second.downpass}
\title{Second downpass}
\usage{
second.downpass(states_matrix, tree)
}
\arguments{
\item{states_matrix}{A \code{list} contains all the states and the activations}

\item{tree}{A \code{phylo} tree}
}
\description{
Applies a second down pass to a node
}
\author{
Thomas Guillerme
}

