% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.nexus.R
\name{str.right}
\alias{str.right}
\title{Rightmost character of string}
\usage{
\method{str}{right}(string, len = nchar(string))
}
\description{
Rightmost character of string
}
\author{
Martin R. Smith
}
\keyword{internal}
