% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.reconstruction.R
\name{apply.reconstruction}
\alias{apply.reconstruction}
\title{Apply reconstruction}
\usage{
apply.reconstruction(tree, character, passes = 4, method, inapplicable,
  match.tip.char = FALSE)
}
\arguments{
\item{tree}{\code{phylo}, a tree}

\item{character}{\code{character}, a vector of character states}

\item{passes}{\code{numeric}, the number of passes in the tree; from \code{1} to \code{4} (default)}

\item{method}{either \code{"Fitch"} or \code{"NA"}}

\item{inapplicable}{When method is \code{"Fitch"}, how do deal with inapplicable data: \code{1}, \code{2} for respectively treating them as ? or an extra state.}

\item{match.tip.char}{\code{logical}, \code{TRUE} to match the character to the tip labels (e.g character 1 matches with tip "a" or "1") or \code{FALSE} (default) to match the character to the tips entry (e.g. character 1 matches with the first tip)}
}
\description{
Apply a series of passes for an ancestral states reconstruction
}
\author{
Thomas Guillerme
}

