% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA.reconstruction.R
\name{first.uppass}
\alias{first.uppass}
\title{First uppass}
\usage{
first.uppass(states_matrix, tree)
}
\arguments{
\item{states_matrix}{A \code{list} contains all the states and the activations}

\item{tree}{A \code{phylo} tree}
}
\description{
Applies a first uppass pass to a node
}
\author{
Thomas Guillerme
}

