% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fitch.reconstruction.R
\name{fitch.downpass}
\alias{fitch.downpass}
\title{Fitch downpass}
\usage{
fitch.downpass(states_matrix)
}
\arguments{
\item{states_matrix}{A \code{list} contains all the states and the activations}
}
\description{
Applies a Fitch down pass to a node
}
\examples{
## Create a states matrix for reconstruction
tree <- ape::read.tree(text = "((a,b),(c,d));")
matrix <- make.states.matrix(tree, "01?-", inapplicable = 1)

## Apply the fitch downpass
(matrix <- fitch.downpass(matrix))

## Access the states
matrix$Dp1

}
\seealso{
\code{\link{apply.reconstruction}}
}
\author{
Thomas Guillerme
}
