% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fitch.reconstruction.R
\name{fitch.uppass}
\alias{fitch.uppass}
\title{Fitch uppass}
\usage{
fitch.uppass(states_matrix)
}
\arguments{
\item{states_matrix}{A \code{list} contains all the states and the activations}
}
\description{
Applies a Fitch up pass to a node
}
\author{
Thomas Guillerme
}

