% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inapp-package.R
\docType{package}
\name{Inapp-package}
\alias{Inapp-package}
\alias{Inapp}
\title{Inapplicable Discrete Characters Reconstruction}
\description{
An algorithm allowing states reconstruction and tree score calculation for inapplicable discrete characters
}
\examples{
\dontrun{
runInapp()
}

}
\references{
Brazeau, M. D., Guillerme, T., and Smith, M. R. (2019). An algorithm for
Morphological Phylogenetic Analysis with Inapplicable Data.
Systematic Biology 68(4): 619-631. \url{https://dx.doi.org/10.1093/sysbio/syy083}
}
\author{
Thomas Guillerme \href{mailto:guillert@tcd.ie}{guillert@tcd.ie} & Martin D. Brazeau & Martin R. Smith.
}
\keyword{characters,}
\keyword{inapplicable}
\keyword{maximum}
\keyword{parsimony,}
\keyword{reconstruction,}
\keyword{score}
\keyword{states}
\keyword{tree}
