% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inapp-package.R
\docType{package}
\name{Inapp-package}
\alias{Inapp-package}
\alias{Inapp}
\title{Inapplicable Discrete Characters Reconstruction}
\description{
An algorithm allowing states reconstruction and tree score calculation for inapplicable discrete characters
}
\examples{
\dontrun{
runInapp()
}

}
\references{
Brazeau, M. D., Guillerme, T., and Smith, M. R. 2017. Morphological 
phylogenetic analysis with inapplicable data. Biorxiv. <https://doi.org/10.1101/209775>
}
\author{
Thomas Guillerme <guillert@tcd.ie>
}
\keyword{characters,}
\keyword{inapplicable}
\keyword{maximum}
\keyword{parsimony,}
\keyword{reconstruction,}
\keyword{score}
\keyword{states}
\keyword{tree}
