% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA.utilities.R
\name{make.states.matrix}
\alias{make.states.matrix}
\title{Sates matrix}
\usage{
make.states.matrix(tree, character, inapplicable = NULL,
  match.tip.char = FALSE)
}
\arguments{
\item{tree}{\code{phylo}, a tree}

\item{character}{Either vector of character states (\code{"numeric"} or \code{"character"}) or a list of the same length of than the tips in the tree (see details)}

\item{match.tip.char}{\code{logical}, \code{TRUE} to match the character to the tip labels (e.g character 1 matches with tip "a" or "1") or \code{FALSE} (default) to match the character to the tips entry (e.g. character 1 matches with the first tip)}
}
\description{
Creates a states matrix
}
\details{
If \code{character} argument is a list, each element of the list must be a \code{"numeric"} vector with \code{"?"} being all states and \code{"-"} being \code{-1}.
}
\author{
Thomas Guillerme
}

