% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg.R
\name{PrintJavascript}
\alias{PrintJavascript}
\title{Print Javascript}
\usage{
PrintJavascript(filepath)
}
\arguments{
\item{filepath}{Path to Javascript template}
}
\value{
Prints to stdout the javascript file, executing any R code included
in `r ...`.
}
\description{
Print Javascript
}
\author{
Martin R. Smith
}
