% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.nexus.R
\name{read.characters}
\alias{read.characters}
\title{Read characters from Nexus file}
\usage{
read.characters(filepath)
}
\arguments{
\item{filepath}{character string specifying location of file}
}
\value{
Character data
}
\description{
Parses Nexus file, reading character states and names
}
\details{
Tested with nexus files downloaded from MorphoBank with the "no notes"
option, but should also work more generally.

Do (report)\link{https://github.org/TGuillerme/Inapp/issues} incorrectly parsed files.
}
\author{
Martin R. Smith
}
