% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runInapp.R
\name{runInapp}
\alias{runInapp}
\title{Run Inapp}
\usage{
runInapp(...)
}
\arguments{
\item{...}{Any optional option to be passed to shiny::runApp}
}
\description{
Running the Inapp GUI via shiny
}
\examples{
\dontrun{
## Running the GUI on your web browser
runInapp()
}

}
\author{
Thomas Guillerme
}
\seealso{
\code{\link{apply.reconstruction}}, \code{\link{plot.states.matrix}}
}

