% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg.R
\name{SVGCanvas}
\alias{SVGCanvas}
\title{SVG canvas}
\usage{
SVGCanvas(trees, outgroupTips = NULL, analysisNames = character(0),
  width = 682, height = 682, uniqueTrees = unique(trees),
  xMargins = c(10L, 10L), yMargins = c(10L, 10L), textY = 3L)
}
\arguments{
\item{trees}{A list of trees, perhaps of class multiPhylo}

\item{outgroupTips}{Character vector specifying tips to include in the outgroup,
to be sent to \link[TreeSearch:RootTree]{TreeSearch::RootTree}.}

\item{analysisNames}{Character vector of length \code{length(trees)}, to be used to
specify the analytical method under which each tree in \code{trees} was generated
(e.g. equal weights parsimony, maximum likelihood).  Will be displayed as a
legend beside each tree.  A list of methods will be displayed beside an output
tree as a legend; analyses associated with a tree will be given the CSS class
\code{thisAn}; other analyses will bear the class \code{notThisAn}.
Defaults to \code{character(0)}, which suppresses this legend.}

\item{width, height}{Numeric specifying the size of the plot in pixels}

\item{uniqueTrees}{(optional) A list generated by unique(trees).  Can be
prepopulated to avoid time-intensive recalculation.}

\item{xMargins, yMargins}{Numeric vectors of length 2 specifying, in pixels,
the margin to be left at each side of the plot}

\item{textY}{Numeric specifying how far to adjust the vertical position of text}
}
\value{
An object of class SVGCanvas
}
\description{
Creates an object of class \code{SVGCanvas}.
}
\details{
\code{SVGCanvas} objects are used to visualize character reconstructions on
phylogenetic trees.
The vizualizations are rendered as SVG elements that can be saved
as stand-alone SVG files, or included in HTML documents.

If incorporated in an HTML vignette, Javascript and CSS can be used to
style the output and enable users to interact with the visualizations.
}
\author{
Martin R. Smith
}
