% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg.R
\name{SVGCanvas}
\alias{SVGCanvas}
\title{SVG canvas}
\usage{
SVGCanvas(trees, outgroupTips, analysisNames = character(0), width = 682,
  height = 682, xMargins = c(10L, 10L), yMargins = c(10L, 10L))
}
\arguments{
\item{trees}{A list of trees, perhaps of class multiPhylo}

\item{outgroupTips}{Character vector specifying tips to include in the outgroup,
to be sent to [TreeSearch::RootTree].}

\item{analysisNames}{Character vector specifying name of each analysis, or
\code{character(0)} if not to be shown.}

\item{width, height}{Numeric specifying the size of the plot in pixels}

\item{xMargins, yMargins}{Numeric vectors of length 2 specifying the margin in pixels
to be left at each side of the plot}
}
\value{
An object of class SVGCanvas
}
\description{
SVG canvas
}
\author{
Martin R. Smith
}
