% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignettePlots.R
\name{vignettePlot}
\alias{vignettePlot}
\title{Vignette Plot}
\usage{
vignettePlot(tree, character, na = TRUE, legend.pos = "bottomleft",
  passes = if (na) 1:4 else 1:2, state.labels = character(0),
  counts = if (na) 1:2 else 2, state.override = NULL,
  score.override = NULL, changes.override = NULL,
  regions.override = NULL, ...)
}
\arguments{
\item{tree}{Tree to plot}

\item{character}{\code{character} string listing character states to plot;
tokens will be matched to labels in alphabetical order.}

\item{na}{\code{logical}, should \code{-} be treated as \code{inapplicable}?}

\item{legend.pos}{\code{character} Position to plot legend.}

\item{passes}{\code{integer} specifying which passes to plot; specify 0 for none.}

\item{state.labels}{Will be passed directly to \code{\link{plot.states.matrix}}}

\item{counts}{\code{integer} vector specifying which counts to show: see \link{plot.states.matrix} for details}

\item{state.override}{\code{list} specifying states to reconstruct at each tip
and node, instead of those reconstructed}

\item{score.override, changes.override, regions.override}{\code{integer}s to
override the values calculated}

\item{\dots}{Additional parameters to pass to \code{\link{plot.states.matrix}}.}
}
\description{
Creates a plot formatted for inclusion in vignettes
}
\author{
Martin R. Smith
}
\keyword{Internal}
